Combine asset helpers
